Sane way of adding many pieces of metadata
