+++ modules/lo_event/examples/browser_events.html
@@ -12,6 +12,26 @@
 <div id="output"></div>
 <script type="module">
   import { subscribeToEvents } from '../lo_event/browser_events.js';
+  import * as lo_event from '../lo_event/lo_event.js';
+  import { consoleLogger } from '../lo_event/consoleLogger.js';
+  import { browserInfo } from '../lo_event/metadata/browserinfo.js';
+  import { chromeAuth } from  '../lo_event/metadata/chromeauth.js';
+  import { localStorageInfo, sessionStorageInfo } from  '../lo_event/metadata/storage.js';
+
+  lo_event.init (
+      "org.ets.browser-test",
+      "0.0.1",
+      [consoleLogger()],
+      {
+          metadata: [
+              browserInfo(),
+              chromeAuth(),
+              localStorageInfo(),
+              sessionStorageInfo()
+          ]
+      }
+  );
+  lo_event.go();
 
   function logToDiv(eventType, event) {
       // Select the output div
