+++ modules/lo_event/lo_event/util.js
@@ -24,147 +24,6 @@ export function copyFields (source, fields) {
 }
 
 /*
-  Browser information object, primarily for debugging. Note that not
-  all fields will be available in all browsers and contexts. If not
-  available, it will return null (this is even usable in node.js,
-  but it will simply return that there is no navigator, window, or
-  document object).
-
-  @returns {Object} An object containing the browser's navigator, window, and document information.
-
-  Example usage:
-    const browserInfo = getBrowserInfo();
-    console.log(browserInfo);
-*/
-export function getBrowserInfo () {
-  const fields = [
-    'appCodeName',
-    'appName',
-    'buildID',
-    'cookieEnabled',
-    'deviceMemory',
-    'language',
-    'languages',
-    'onLine',
-    'oscpu',
-    'platform',
-    'product',
-    'productSub',
-    'userAgent',
-    'webdriver'
-  ];
-
-  const connectionFields = [
-    'effectiveType',
-    'rtt',
-    'downlink',
-    'type',
-    'downlinkMax'
-  ];
-
-  const documentFields = [
-    'URL',
-    'baseURI',
-    'characterSet',
-    'charset',
-    'compatMode',
-    'cookie',
-    'currentScript',
-    'designMode',
-    'dir',
-    'doctype',
-    'documentURI',
-    'domain',
-    'fullscreen',
-    'fullscreenEnabled',
-    'hidden',
-    'inputEncoding',
-    'isConnected',
-    'lastModified',
-    'location',
-    'mozSyntheticDocument',
-    'pictureInPictureEnabled',
-    'plugins',
-    'readyState',
-    'referrer',
-    'title',
-    'visibilityState'
-  ];
-
-  const windowFields = [
-    'closed',
-    'defaultStatus',
-    'innerHeight',
-    'innerWidth',
-    'name',
-    'outerHeight',
-    'outerWidth',
-    'pageXOffset',
-    'pageYOffset',
-    'screenX',
-    'screenY',
-    'status'
-  ];
-
-  const browserInfo = {
-    navigator: typeof navigator !== 'undefined' ? copyFields(navigator, fields) : null,
-    connection: typeof navigator !== 'undefined' && navigator.connection ? copyFields(navigator.connection, connectionFields) : null,
-    document: typeof document !== 'undefined' ? copyFields(document, documentFields) : null,
-    window: typeof window !== 'undefined' ? copyFields(window, windowFields) : null
-  };
-
-  return { browser_info: browserInfo };
-}
-
-/*
-  This function is a wrapper for retrieving profile information using
-  the Chrome browser's identity API. It addresses a bug in the Chrome
-  function and converts it into a modern async function. The bug it
-  works around can be found at
-  https://bugs.chromium.org/p/chromium/issues/detail?id=907425#c6.
-
-  To do: Add chrome.identity.getAuthToken() to retrieve an
-  authentication token, so we can do real authentication.
-
-  Returns:
-    A Promise that resolves with the user's profile information.
-
-  Example usage:
-    const profileInfo = await profileInfoWrapper();
-    console.log(profileInfo);
-
-  Users should switch to chromeIdentityHeader(), below. This function name should also
-  mention `chrome`, but I don't want to do that without updating the extension at the
-  same time.
-*/
-export function profileInfoWrapper () {
-  if (typeof chrome !== 'undefined' && chrome.identity) {
-    try {
-      return new Promise((resolve, reject) => {
-        chrome.identity.getProfileUserInfo({ accountStatus: 'ANY' }, function (data) {
-          resolve(data);
-        });
-      });
-    } catch (e) {
-      return new Promise((resolve, reject) => {
-        chrome.identity.getProfileUserInfo(function (data) {
-          resolve(data);
-        });
-      });
-    }
-  }
-  // Default to an empty object
-  return new Promise((resolve, reject) => {
-    resolve({});
-  });
-}
-
-export async function getChromeHeader() {
-    let result = await profileInfoWrapper();
-    return { event: 'chrome_identity', chrome_identity: result };
-}
-
-/*
   Generate a unique key, which can be used for session IDs, anonymous user IDs, and other similar purposes.
 
   Parameters:
